+++ test/unit/dropdown-spec.js
@@ -57,19 +57,117 @@ describe('Dropdown', function () {
       assert.strictEqual(subject(), dropdown);
     });
 
-    it('should change #shown from false to true', function () {
-      dropdown = new Dropdown({});
-      dropdown.shown = false;
-      subject();
-      assert(dropdown.shown);
+    context('when search results are given', function () {
+      it('should append dropdown items with the search results', function () {
+        dropdown = new Dropdown({});
+        subject();
+        assert.equal(dropdown.items.length, 1);
+        assert(dropdown.items[0] instanceof DropdownItem);
+        assert.equal(dropdown.items[0].searchResult, searchResult);
+      });
+
+      context('and it has not been shown', function () {
+        it('should change #shown from false to true', function () {
+          dropdown = new Dropdown({});
+          dropdown.shown = false;
+          subject();
+          assert(dropdown.shown);
+        });
+
+        ['show', 'shown', 'rendered'].forEach(eventName => {
+          it(`should emit ${eventName} event`, function () {
+            var spy = this.sinon.spy();
+            dropdown = new Dropdown({});
+            dropdown.shown = false;
+            dropdown.on(eventName, spy);
+            subject();
+            assert(spy.calledOnce);
+          });
+        });
+
+        ['hide', 'hidden'].forEach(eventName => {
+          it(`should not emit ${eventName} event`, function () {
+            var spy = this.sinon.spy();
+            dropdown = new Dropdown({});
+            dropdown.shown = true;
+            dropdown.on(eventName, spy);
+            subject();
+            assert(!spy.called);
+          });
+        });
+      });
+
+      context('and it has been shown', function () {
+        it('should emit rendered event', function () {
+          var spy = this.sinon.spy();
+          dropdown = new Dropdown({});
+          dropdown.shown = false;
+          dropdown.on('rendered', spy);
+          subject();
+          assert(spy.calledOnce);
+        });
+
+        ['show', 'shown', 'hide', 'hidden'].forEach(eventName => {
+          it(`should not emit ${eventName} event`, function () {
+            var spy = this.sinon.spy();
+            dropdown = new Dropdown({});
+            dropdown.shown = true;
+            dropdown.on(eventName, spy);
+            subject();
+            assert(!spy.called);
+          });
+        });
+      });
     });
 
-    it('should append dropdown items with the search results', function () {
-      dropdown = new Dropdown({});
-      subject();
-      assert.equal(dropdown.items.length, 1);
-      assert(dropdown.items[0] instanceof DropdownItem);
-      assert.equal(dropdown.items[0].searchResult, searchResult);
+    context('when search results are empty', function () {
+      function subject_() {
+        return dropdown.render([], { top: 0, left: 0 });
+      }
+
+      context('and it has been shown', function () {
+        it('should change #shown from true to false', function () {
+          dropdown = new Dropdown({});
+          dropdown.shown = true;
+          subject_();
+          assert(!dropdown.shown);
+        });
+
+        ['hide', 'hidden'].forEach(eventName => {
+          it(`should emit ${eventName} event`, function () {
+            var spy = this.sinon.spy();
+            dropdown = new Dropdown({});
+            dropdown.shown = true;
+            dropdown.on(eventName, spy);
+            subject_();
+            assert(spy.calledOnce);
+          });
+        });
+
+        ['show', 'shown', 'rendered'].forEach(eventName => {
+          it(`should not emit ${eventName} event`, function () {
+            var spy = this.sinon.spy();
+            dropdown = new Dropdown({});
+            dropdown.shown = true;
+            dropdown.on(eventName, spy);
+            subject_();
+            assert(!spy.called);
+          });
+        });
+      });
+
+      context('and it has not been shown', function () {
+        ['show', 'shown', 'rendered', 'hide', 'hidden'].forEach(eventName => {
+          it(`should not emit ${eventName} event`, function () {
+            var spy = this.sinon.spy();
+            dropdown = new Dropdown({});
+            dropdown.shown = false;
+            dropdown.on(eventName, spy);
+            subject_();
+            assert(!spy.called);
+          });
+        });
+      });
     });
 
     context('when header option is given', function () {