+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 - Enable to use markdown in jsdoc.
 - Add `cache` strategy parameter.
 - Add `className`, `style`, `maxCount`, `header` and `footer` dropdown options.
+- Add `show`, `shown`, `rendered`, `hide` and `hidden` events to `Textcomplete`.
 
 ### Changed
 - Exclude src/doc from Inch CI.